+++ internal/config/crd/pool.yaml
