+++ internal/config/crd/pool_snapshot.yaml
